Controls growing rate of the abstract test class
