+++ tests/Doctrine/Tests/Models/CaseSensitiveDiscriminatorMap/Cube.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\Tests\Models\CaseSensitiveDiscriminatorMap;
+
+/** @Entity */
+final class Cube extends Shape
+{
+}
