+++ tests/Doctrine/Tests/ORM/Mapping/xml/Doctrine.Tests.Models.CaseSensitiveDiscriminatorMap.Shape.dcm.xml
@@ -3,10 +3,10 @@
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                           https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <entity name="Doctrine\Tests\ORM\Mapping\Shape" inheritance-type="SINGLE_TABLE">
+    <entity name="Doctrine\Tests\Models\CaseSensitiveDiscriminatorMap\Shape" inheritance-type="SINGLE_TABLE">
         <discriminator-column name="discr" type="string" length="32" />
         <discriminator-map>
-            <discriminator-mapping value="cube" class="Doctrine\Tests\ORM\Mapping\cube" />
+            <discriminator-mapping value="cube" class="Doctrine\Tests\Models\CaseSensitiveDiscriminatorMap\cube" />
         </discriminator-map>
         <id name="id" type="integer" column="id">
             <generator strategy="AUTO" />
